requirements.txt
fastapi==0.95.0
uvicorn[standard]==0.23.2
paho-mqtt==1.6.1
jinja2==3.1.2
